+++ Kwetter/src/main/java/service/UserServiceImpl.java
@@ -29,7 +29,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User findUserById(int id) {
+    public User findUserById(long id) {
         return userDAO.GetUserById(id);
     }
